refactor: move parts of the report runner to be more local to the view report route
