+++ src/routes/_auth/reports/$reportId/-runner/presentation/json-view.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { useReportContext } from "@/lib/context/view-report";
+import { useReportContext } from "@/routes/_auth/reports/$reportId/-runner/view-report-context";
 
 const JsonView = () => {
   const { resultState } = useReportContext();
